+++ lib/typed_operation.rb
@@ -1,5 +1,12 @@
+if Gem::Version.new(RUBY_VERSION) < Gem::Version.new("3.2.0")
+  require "polyfill-data"
+end
+
 require "typed_operation/version"
-require "typed_operation/railtie"
+require "typed_operation/railtie" if defined?(Rails::Railtie)
+require "typed_operation/nilable_type"
+require "typed_operation/attribute_builder"
+require "typed_operation/curried"
 require "typed_operation/base"
 require "typed_operation/partially_applied"
 require "typed_operation/prepared"