+++ lib/typed_operation/base.rb
@@ -1,43 +1,88 @@
 # frozen_string_literal: true
 
-require "vident/typed"
-require "vident/typed/attributes"
+require "literal"
 
 module TypedOperation
-  class Base
-    include Vident::Typed::Attributes
-
+  class Base < Literal::Data
     class << self
       def call(...)
         new(...).call
       end
 
-      def curry(**args)
-        PartiallyApplied.new(self, **args).curry
+      def with(...)
+        PartiallyApplied.new(self, ...).with
       end
-      alias_method :[], :curry
-      alias_method :with, :curry
+      alias_method :[], :with
+
+      def curry
+        Curried.new(self)
+      end
 
       def to_proc
         method(:call).to_proc
       end
 
-      def operation_key
-        name.underscore.to_sym
+      # Method to define parameters for your operation.
+
+      # Parameter for keyword argument, or a positional argument if you use positional: true
+      # Required, but you can set a default or use optional: true if you want optional
+      def param(name, signature = :any, **options, &converter)
+        AttributeBuilder.new(self, name, signature, options).define(&converter)
       end
 
-      # property are required by default, you can fall back to attribute or set allow_nil: true if you want optional
-      def param(name, signature = :any, **options, &converter)
-        attribute(name, signature, **{allow_nil: false}.merge(options), &converter)
+      # Alternative DSL
+
+      # Parameter for positional argument
+      def positional(name, signature = :any, **options, &converter)
+        param(name, signature, **options.merge(positional: true), &converter)
+      end
+
+      # Parameter for a keyword or named argument
+      def named(name, signature = :any, **options, &converter)
+        param(name, signature, **options.merge(positional: false), &converter)
+      end
+
+      # Wrap a type signature in a NilableType meaning it is optional to TypedOperation
+      def optional(type_signature)
+        NilableType.new(type_signature)
+      end
+
+      # Introspection methods
+
+      def positional_parameters
+        literal_attributes.filter_map { |name, attribute| name if attribute.positional? }
+      end
+
+      def keyword_parameters
+        literal_attributes.filter_map { |name, attribute| name unless attribute.positional? }
+      end
+
+      def required_positional_parameters
+        required_parameters.filter_map { |name, attribute| name if attribute.positional? }
+      end
+
+      def required_keyword_parameters
+        required_parameters.filter_map { |name, attribute| name unless attribute.positional? }
+      end
+
+      def optional_positional_parameters
+        positional_parameters - required_positional_parameters
+      end
+
+      def optional_keyword_parameters
+        keyword_parameters - required_keyword_parameters
+      end
+
+      private
+
+      def required_parameters
+        literal_attributes.filter do |name, attribute|
+          attribute.default.nil? # Any optional parameters will have a default value/proc in their Literal::Attribute
+        end
       end
     end
 
-    def initialize(**attributes)
-      begin
-        prepare_attributes(attributes)
-      rescue ::Dry::Struct::Error => e
-        raise ParameterError, e.message
-      end
+    def after_initialization
       prepare if respond_to?(:prepare)
     end
 
@@ -53,14 +98,9 @@ module TypedOperation
       attributes.values
     end
 
-    def deconstruct_keys(_keys)
-      attributes.to_h
-    end
-
-    private
-
-    def operation_key
-      self.class.operation_key
+    def deconstruct_keys(keys)
+      h = attributes.to_h
+      keys ? h.slice(*keys) : h
     end
   end
 end