+++ lib/typed_operation/nilable_type.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require "literal"
+
+module TypedOperation
+  class NilableType < Literal::Union
+    def initialize(*types)
+      @types = types + [NilClass]
+    end
+  end
+end