+++ lib/typed_operation/prepared.rb
@@ -3,7 +3,11 @@
 module TypedOperation
   class Prepared < PartiallyApplied
     def operation
-      @operation.new(**@applied_args)
+      operation_class.new(*@positional_args, **@keyword_args)
+    end
+
+    def prepared?
+      true
     end
   end
 end